b=24141 llite: -EIO instead of LBUG for multi-referenced object
Whenever an inode is used with a DLM lock, the client checks that no
other inodes are referencing the same OST object, since this is a sign
of filesystem corruption on the MDS (or some other code bug that
behaves in this way).  If the client detected the same OST object is
referenced from multiple inodes at the same time, it will LASSERT()
and print a message to this effect, rather than continue to corrupt
the data files.

  osc_set_data_with_check() ASSERTION(old_inode->i_state & I_FREEING)
  failed: Found existing inode ffff880587d15d10/222311317/67781718
  state 0 in lock: setting data to ffff88046b7f8d50/223489633/67781099

Instead of LASSERTing on this condition, instead return EIO for this
file.  This allows the problem to be analyzed and fixed without the
need to reboot the client node.

i=adilger