b=24450 ptlrpc_set_wait flexibility
ptlrpc_set_wait() waits until all requests in a set complete.
This patch makes it possible to specify a condition on which ptlrpc_set_wait()
will wait instead of default condition "no remaining requests"

With that it wiil be possible to add requests to a set as sent ones complete
without waiting for all requests to finish.

i=nathan.rutman