LU-1053 build: allow deleted/chmod in commit-msg
Allow files that were deleted or chmod when verifying a commit
message that contains a diff generated with "commit -v".  The
checks added for excluding the diff were previously too strict.
Add a test case for each of these situations, as well as a case
for a commit message that contains "diff" at the start of a line
in the middle of a commit message.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I74fe98682342e445da7fe64fd42e91df4564500c
Reviewed-on: http://review.whamcloud.com/2053
Reviewed-by: Bruce Korb <bruce_korb@xyratex.com>
Reviewed-by: James Simmons <uja.ornl@gmail.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>