LU-1160 tests: remove liblustre from default tests
The liblustre testing is currently failing. Since this functionality
is not being used by any (or at least not many?) users since it is
no longer used for the Cray Catamount compute clients, the liblustre
testing should be removed from automated test runs.

Until the test can be fixed this test can be re-enabled. As yet,
the liblustre code will not be removed from the tree.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I916994b71a60e0f8596b7babb69c67a8b2bebc73
Reviewed-on: http://review.whamcloud.com/2244
Tested-by: Hudson
Reviewed-by: Minh Diep <mdiep@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>