MRP-153 Add "--erase-param", change "--param" behaviour in tunefs.lustre
A new option is added to tunefs.lustre, "--erase-param", which allows to
remove all instances of a specific parameter.

This patch should come with a warning, as it introduces a change in current
behaviour: option "--param" is modified to erase all previously stored
instances of the parameter, and store all new instances of the parameter
specified in the command line.

Signed-off-by: Nikitas Angelinas <nikitas_angelinas@xyratex.com>
Reviewed-by: Vitaly Fertman <vitaly_fertman@xyratex.com>
Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>