MRP-397 mgs: "replace_nids" lctl command added
Currently after device address is changed tunefs should be started with
"writeconf" options. This erase configuration logs for the filesystem
that this server is part of, and regenerate them. This is very dangerous.
All clients must be unmounted and servers  for  this filesystem should be
stopped.  All targets (OSTs/MDTs) must then be restarted to regenerate
the logs.  No clients should be started until all targets have restarted.
The targets start order is also important. Wrong order can be cause of
failure.

This patch adds "replace_nids" command to lctl that should be executed then
only MGS server started. This command replace nids and uuid for given
device. Command can be used to change OST or MDT nids.

Reviewed-by: Nathan Rutman <nathan_rutman@xyratex.com>
Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>