MRP-542 DIO sends wrong data on replay.
In a case of DIO, RPC has a DIRECT mapping to a userland buffer. after no
transno flag set - target stops to send a last commited transaction number,
so network client thinks the request completed OK, but data not on disk and
move RPC to replay queue (request transno > last know committed transno).
next operation refills buffers with different data but RPC is still in the
replay queue with the mapping to these buffers, i.e. wrong date are sent on
replay.

Signed-off-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>
Reviewed-by: Alexander Zarochentsev <Alexander_Zarochentsev@xyratex.com>
Reviewed-by: Andrew Perepechko <andrew_perepechko@xyratex.com>