MRP-678 kernel: Fix complication issue on 32 bit build
When compiling the kernel on a 32 bit build the following
compilation error was generated:
error: #error SECTIONS_WIDTH+NODES_WIDTH+ZONES_WIDTH
> BITS_PER_LONG - NR_PAGEFLAGS
This is because the patches for raid5_zero_copy add a
PG_constant enum to the list, which increases NR_PAGEFLAGS
by one which results in this condition being hit. The solution is to
add a configuration option that will allow 64 bit build from using
this constant, and 32 bit builds to not use it. Since we are not too
concerned with 32 bit server this should not be an issue. If the
value is not present the zero copy portion of the change
will not be activated.

WC-bug-id: LU-535

Signed-off-by: Michael Mckay <michael_mckay@xyratex.com>
Change-Id: Ida468030e6fc2dd561d091b30c36bd759fbad5ff
Reviewed-on: https://morpheus.xyus.xyratex.com:8443/gerrit/103
Tested-by: Jenkins
Reviewed-by: Bruce Korb <Bruce_Korb@xyratex.com>
Reviewed-by: Eric Mei <Eric_Mei@xyratex.com>
Reviewed-by: Roman Grigoryev <Roman_Grigoryev@xyratex.com>
Tested-by: Alexander Lezhoev <Alexander_Lezhoev@xyratex.com>
Reviewed-by: Vitaly Fertman <Vitaly_Fertman@xyratex.com>