LU-107 scripts: rename heartbeat resource agent
Add a .ha_v2 extension to the Lustre heartbeat resource agent script
to resolve conflicts with the lustre init script on case-insensitive
filesystems.  The extension also provides a clue as to what the
script's purpose is.  Add a compatibility symlink with the plain name
during package installation.   If we later add v3-style scripts we can
control which one gets linked to via a configure option.

Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: I0e37a390fbfb3f00c1c1e666a7cdbe5d37fa885b
Reviewed-on: http://review.whamcloud.com/4233
Tested-by: Hudson
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>