MRP-459 ptlrpc: Allow clients reconnect with RPCs in process
Since not letting clients with pending RPCs to reconnect caused a lot
of problems, this patch allows such a reconnection and would only
refuse to process requests with the same xids that are already being
processed. This patch is a port of LU-793 to b_neo_stable (2.2)
Requests from Andreas were added also.

Signed-off-by: Iurii Golovach <iurii_golovach@xyratex.com>
Signed-off-by: Andriy Skulysh <Andriy_Skulysh@xyratex.com>
Change-Id: I954e43aa0242304707d6b0600df4ad229fb47227
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/231
Tested-by: Jenkins
Reviewed-by: Alexey Lyashkov <Alexey_Lyashkov@xyratex.com>
Tested-by: Alexander Lezhoev <Alexander_Lezhoev@xyratex.com>
Reviewed-by: Denis Kondratenko <Denis_Kondratenko@xyratex.com>