MRP-814 tests: combine dependent sanity tests
Some tests can not run as expected when previous dependent test is not
ran.
This patch combines the following sanity tests:
  tests 1a, 1b to test 1, test 4 is removed as duplicate of new test 1
  tests 2a, 2b to test 2, test 3 is removed as duplicate of new test 2
  tests 3a, 3b, 3c to test 3
  tests 4a, 4b to test 4
  tests 6a, 6b to test 6a
  tests 6c, 6d to test 6c
  tests 6e, 6f to test 6e
For the changed tests $tfile and $tdir are used for test file and dir
names.
Error exit messages are added to indicate what went wrong.
Test 0 is renamed to test 0a to allow running it separately.

Signed-off-by: Vladimir Saveliev <vladimir_saveliev@xyratex.com>
Change-Id: If79acff4e4eec948ee42382c90c9a8f60bb3ce03
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/301
Tested-by: Jenkins
Reviewed-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Reviewed-by: Kyrylo Shatskyy <kyrylo_shatskyy@xyratex.com>
Tested-by: Kyrylo Shatskyy <kyrylo_shatskyy@xyratex.com>
Reviewed-by: Vitaly Fertman <Vitaly_Fertman@xyratex.com>