MRP-989 recovery-small 19a/19b failure on multinode setup.
OBD_FAIL_LDLM_CANCEL_NET (0x304) allows insertion of failure
points into the Lustre code. This needs to be set on servers
instead if clients.

The fix to ldlm_drop_cancel is introduced by LU-1341.

This change enables the fail_loc to be
executed on servers, resulting in the necessary recovery
scenario to test "expired_lock_main on mds/ost" successfully.

WC-bug-id: LU-1341
Signed-off-by: vitaly_fertman@xyratex.com
Signed-off-by: parinay_kondekar@xyratex.com
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/#/c/388/
Change-Id: Iff7429562619a988c0ff07e837ad668d0e1dbcef
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/388
Tested-by: Jenkins
Reviewed-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Reviewed-by: Vitaly Fertman <Vitaly_Fertman@xyratex.com>