MRP-1057 tests: check lustre.conf for modprobe
The newer userspace modprobe requires that the /etc/modprobe.d/ files
end in ".conf", so that there is no confusion between config files
that are supposed to be parsed, and files that are left after editing
(e.g. lustre.conf.orig or lustre.conf~).

Add a check for /etc/modprobe.d/lustre.conf to get lnet module
parameters during testing.

WC-bug-id: LU-506
Signed-off-by: Vladimir Saveliev <vladimir_saveliev@xyratex.com>
Change-Id: I68ca9802e010faec79120f56dc02516f774bbb38
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/446
Tested-by: Jenkins
Reviewed-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>