LU-3720 test: use actual stripe count in sanity.sh 130d
The test is expecting each OST to get a specific amount if data based
in the size and stripe count, but if an OST is full, more data will be
written to each of the other OSTs.
To fix that, use the actual number of stripes allocated instead of the
requested number of stripes.

Signed-off-by: Emoly Liu <emoly.liu@intel.com>
Change-Id: Iadb0daf16d7c121d2104f1fe49afa7f762a7f52a
Reviewed-on: http://review.whamcloud.com/7303
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: James Nunez <james.a.nunez@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>