LU-3478 iokit: fix sgpdd-survey scripts (output and plotting)
The sgpdd-survey script does not provide minimal and maximal bandwidth
of sgp_dd instances. This is important to be able to identify if low
LUNs are present. This patch updates the output to display this
information, similarly to obdfilter-survey.

The plot-sgpdd script is updated to handle this output format
change. It is also reorganized to simplify operation handling.

Signed-off-by: Gregoire Pichon <gregoire.pichon@bull.net>
Change-Id: Ie6ad57951dff8c8b2871457652a94df6dccbc491
Reviewed-on: http://review.whamcloud.com/6680
Reviewed-by: Minh Diep <minh.diep@intel.com>
Reviewed-by: Bobbie Lind <bobbie.j.lind@intel.com>
Reviewed-by: Cliff White <cliff.white@intel.com>
Tested-by: Cliff White <cliff.white@intel.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>