LU-1175 tests: use ofd by default instead of obdfilter
XXX: partially ported. OFD is not in use yet

There was a bug in comma_list() appending a space to the list of
nodes, which caused do_rpc_nodes() to be unhappy when a nearly-empty
list of nodes was being passed by callers.  Rework this function, and
clean up other callers to just pass the nodes and let do_rpc_nodes()
detect this case and skip any remote execution.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I896c9c9b7216c418b635e893202810333d8ef871
Reviewed-on: http://review.whamcloud.com/4171
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Mike Pershin <tappro@whamcloud.com>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>