MRP-459 ptlrpc: Allow clients reconnect with RPCs in process
Since not letting clients with pending RPCs to reconnect caused a lot
of problems, this patch allows such a reconnection and would only
refuse to process requests with the same xids that are already being
processed. This patch is a port of LU-793 to the b_neo_stable 1.3
Requests from Andreas were added.

Signed-off-by: Iurii Golovach <iurii_golovach@xyratex.com>
Signed-off-by: Andriy Skulysh <Andriy_Skulysh@xyratex.com>
Signed-off-by: Alexander Boyko <alexander_boyko@xyratex.com>
Reviewed-by: Alexander Zarochentsev <alexander_zarochentsev@xyratex.com>
Reviewed-by: Denis Kondratenko <Denis_Kondratenko@xyratex.com>
Change-Id: I1952e793674d40d1a5c77110fab818c907686a40
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/626
Tested-by: Jenkins
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Tested-by: Alexander Lezhoev <Alexander_Lezhoev@xyratex.com>