MRP-1362 tests: Lustre Sanity test_101b hangs w/ OSTCOUNT>100
sanity.sh test_101b hangs during Lustre sanity test. This
is happening especially in cases where the $OSTCOUNT is
greater than 100.

Due to OSTCOUNT of more than 100, the ITERATION parameter
results in '0', which causes the 'reads' utility to run
in infinite 'for' loop. Thus resulting in a hang on such
setups.

This change would make necessary changes in FILE_LENGTH in
test_101b and its related routines(setup_101bc, ra_check)
so that, the test won't hang and would proceed rightly.

Author: Parinay Kondekar <Parinay_Kondekar@xyratex.com>

Signed-off-by: Parinay Kondekar <Parinay_Kondekar@xyratex.com>
Change-Id: Ida4b54f5129c9577dbb966289d7d18aec7071f0c
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/715
Tested-by: Jenkins
Reviewed-by: Alexander Boyko <Alexander_Boyko@xyratex.com>
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Reviewed-by: Vladimir Saveliev <vladimir_saveliev@xyratex.com>
Reviewed-by: Vitaly Fertman <Vitaly_Fertman@xyratex.com>