MRP-1263 llite: lustre writepages address space operation
When a client generates dirty pages faster than servers commit them to
disk, it may lead to OOM. Pages sent to servers are not marked dirty.
That confuses write throttling so that it allows to generate more
dirty pages when it should throttle a writer.

This patch turns pages sent to server but not yet committed into dirty
memory accounting and implements writepages address space operation
for lustre:
- count those pages as unstable_nfs
- on background writeback: send OST_SYNC rpc to OSTs with pending
  commits
- on foreground writeback: send OST_SYNC rpc to OSTs with pending
  commits and wait for completion. When there is already in flight
  foreground writeback on an OST, wait for its completion as well.
- for "old" data flushing generic_writepages is called only

Signed-off-by: Vladimir Saveliev <vladimir_saveliev@xyratex.com>
Change-Id: I6ef73bd61fda81cbc3da99bae965ddfcbd0e0281
Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/709
Tested-by: Jenkins
Reviewed-by: Alexander Boyko <Alexander_Boyko@xyratex.com>
Tested-by: Elena Gryaznova <Elena_Gryaznova@xyratex.com>
Reviewed-by: Andrew Perepechko <Andrew_Perepechko@xyratex.com>
Reviewed-by: Vitaly Fertman <Vitaly_Fertman@xyratex.com>