LU-3319 procfs: provide framework for seq_file handling
In linux 3.10 kernels and above the proc file system interface
changed. To handle this change we need to move from the current
method of using proc_read and proc_write function pointers to
using the seq_file. This patch provides the base to enable moving
all the lustre proc handling gradually over to this new method
of proc data handling.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I06d941398bf248bffab851f9f25bea6bbe438a30
Reviewed-on: http://review.whamcloud.com/7135
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Yang Sheng <yang.sheng@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>