LU-5041 osd-zfs: Prefetch dnodes
After a successful FID lookup prefetch the referenced dnode.  There's
every reason to believe the dnode will be immediately useful.  This
will not result in a spill block also being prefetched for that we
need the block pointer from the dnode.  However, once large dnodes
are supported spill blocks should be rarely needed.

Change-Id: I41c0d08249f5d852137a08e2f3b262fb664a15ea
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Reviewed-on: http://review.whamcloud.com/10395
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>