LU-4416 osd-ldiskfs: vfs structs use fully typed uid/gid
Since Linux 3.5, struct inode and iattr have used type kuid_t
and kgid_t for uids and gids. There was a compatibility mode
to avoid build failures with older code but it was removed in
3.14. Furthermode, SLES12 builds with UIDGID_STRICT_TYPE_CHECKS
enabled, and needs to use the shim interfaces to build.

This patch provides the shims for older releases without them.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: If4c70b10b4bc13aa90d729d56b25edbe1f52fa98
Reviewed-on: http://review.whamcloud.com/10161
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Peng Tao <bergwolf@gmail.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>