MRP-2041 tests: demonstrate ast error handler flaw
When a client returns an error other than EINVAL replying to blocking
ast, it is unsafe to cancel the lock on server side only, because the
client may continue its I/O assuming it still owns the lock while the
real lock may be granted already to another client.

Simulate situation when lock is cancelled on server but is still valid
on client.

cleanups:
  test_10c renamed to test_65
  test_10d renamed to test_10c
  new test is test_10d
  clearing fail_val is added to reset_fail_loc()

Change-Id: Ief7b8ba619197eee0c8b0096da619b3034c4e23f
Signed-off-by: Vladimir Saveliev <vladimir_saveliev@xyratex.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/3520
Tested-by: Jenkins
Tested-by: Alexander Lezhoev <alexander_lezhoev@xyratex.com>
Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>
Reviewed-by: Vitaly Fertman <vitaly_fertman@xyratex.com>