MRP-1700 obd: lu_object_find_at hung
lu_object_find_at hungs if called 2 times and object is removed
in between. It makes mdt_rename_sanity not workable for rename.
Change mdt_rename_sanity to work on existing object.

the following commit is included:
    LU-5069 mds: check if the object exist before subdir check

also, make it posix compliant, an attempt to make a dir a subdir
of itself must return EINVAL

Signed-off-by: Vitaly Fertman <vitaly_fertman@xyratex.com>
Change-Id: I10dc08de30ae5cd4faaa84eb0d5841ad13267a23