MRP-1989 quota: prevent wait at quota_chk_ack_common
When OST device is deactivated, and osc-mdt is evicted from OST,
ost_brw_write lock the pages and sleep at

[<ffffffffa0ec167e>] cfs_waitq_wait+0xe/0x10 [libcfs]
[<ffffffffa0891260>] quota_chk_acq_common+0x970/0xc50 [lquota]
[<ffffffffa0b6197c>] filter_commitrw_write+0xbac/0x3388 [obdfilter]
[<ffffffffa0b5566f>] filter_commitrw+0x27f/0x2a0 [obdfilter]
[<ffffffffa0b132d6>] obd_commitrw+0x126/0x3d0 [ost]
[<ffffffffa0b1bec4>] ost_brw_write+0x1314/0x1af0 [ost]
[<ffffffffa0b22718>] ost_handle+0x3e38/0x45a0 [ost]

Other ll_ost_io threads which want to lock the same pages,
will hangs too. This wait was designed for MDS failover,
client IO should not be affected in this case. But no
need to wait for quota master infinitely.
Patch adds ability to stop waiting when client reconnected
or failed, and defines max timeout for this wait.

In this case brw resends could not occupy all ost_io threads
and hangs. The script looks like:
1) client send brw
2) ost thread lock pages and wait at quota_chk_acq_common()
3) client see brw request timeout, reconnect and resend brw
4) ost detects client reconnection during wait and
   return -ENOTCONN from quota_chk_ack_common for first brw
5) brw resend repeats wait at quota_chk_ack_common() again

This patch resolves the issue with ll_ost_io hang threads.

Signed-off-by: Alexander Boyko <alexander_boyko@xyratex.com>
Change-Id: I0f8a131d822aa7dea0dfacda8e2726b5eb0d339f
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/3781
Reviewed-by: Andrew Perepechko <andrew_perepechko@xyratex.com>
Tested-by: Jenkins
Reviewed-by: Sergey Cheremencev <sergey_cheremencev@xyratex.com>
Tested-by: Alexander Lezhoev <alexander_lezhoev@xyratex.com>
Reviewed-by: Andriy Skulysh <andriy_skulysh@xyratex.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>