MRP-2145 osc: osc_lock_weight endless loop fix
With huge number of pages to scan by osc_lock_weight() it is likely
CLP_GANG_RESCHED is returned from osc_page_gang_lookup() and the scan will be
repeated again from the start. The fix is to return positive number from
osc_lock_weight() if CLP_GANG_RESCHED occured, considering the lock is heavy
enough to be replayed.

Change-Id: I60775f2d8f688029d97ad4fe64f9b2698ed278c8
Signed-off-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/4007
Tested-by: Jenkins
Reviewed-by: Sergey Cheremencev <sergey_cheremencev@xyratex.com>
Reviewed-by: Andrew Perepechko <andrew_perepechko@xyratex.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Tested-by: Alexander Lezhoev <alexander_lezhoev@xyratex.com>