MRP-1759 build: add kmod for client
When the Lustre is built as client with driver update support,
kmod package has ability to check and copy modules for the new
added kernel. The package includes dependencies on a kernel symbols
version. In most cases this is enough for a minor security kernel
updates. New param build_kmod was added to spec file. For Lustre
client build_kmod is enabled by default, and disabled for servers.

This patch changes the Lustre modules location to
[extra|updates]/lustre-client/[fs|net]/ for client build.
Kmod utils/scripts requires this modification, cause they use %name
for postinstall scripts. Path to Lustre modules for server build
stay the same.

Signed-off-by: Alexander Boyko <alexander.boyko@seagate.com>
Change-Id: I303bd0d4367a48adee61b437f7053a93a2c55c5a
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/4096
Tested-by: Jenkins
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@seagate.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@seagate.com>
Reviewed-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>