MRP-153 utils: add erase-param change param in tunefs.lustre
A new option is added to tunefs.lustre, "--erase-param", which allows
to remove all instances of a specific parameter.

This patch should come with a warning, as it also introduces a change
in current behaviour: option "--param" is modified to erase all
previously stored instances of the parameter, and store all new
instances of the parameter specified in the command line.

Signed-off-by: Nikitas Angelinas <nikitas.angelinas@seagate.com>
Change-Id: Ia981d4b964af07adcd0d9f03be96dcb31711b970
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/4330
Tested-by: Jenkins
Reviewed-by: Artem Blagodarenko <artem.blagodarenko@seagate.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@seagate.com>
Tested-by: Elena Gryaznova <elena.gryaznova@seagate.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>