MRP-2288 mdd: decref volatile object after creation
porting of LU-3696

Drop the nlink count on a volatile file after it is created.
It is created with a single reference and then added to the PENDING
directory, which adds a second reference to the inode.  When it is
closed it is left with an nlink count=1 and not unlinked as it should
be.

Do not insert volatile files into the ChangeLog, since they are
intended to be temporary files that cannot be accessed after creation
and should not be synched to a backup filesystem.

If llapi_create_volatile_idx() is used on a non-Lustre filesystem then
it should also unlink the temporary filename after opening it, so that
the behaviour is the same.  Improve the comments for this function.

Minor cleanups of related functions found during investigation.

Lustre-change: http://review.whamcloud.com/#/c/10179
Lustre-commit: 1ebc8dc473dd9f948837b9d5b258194637d46525

Intel-bug-id: LU-3696

Change-Id: I163dfbeb36610a7c7da911ce501e61b636fd16bb
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Signed-off-by: gaurav mahajan <gaurav.mahajan@seagate.com>