MRP-1690 tests: sanity test_27m failure
sanity 27m fails with "OST0 was full byt new created file still use it"
if the test runs with more than 1 connected client.

The reason is grants. Every client initially gets 2 Mb grant.
When dd from the first client receives ENOSPC, it does not mean
the OST is filled up, since the client is not allowed to use
other clients' grants. When creating a new file, the MDS still
sees free space on OST0 equal to the amount of unused grants
and allocates new objects on OST0.

This situation does not seem to reflect any defect in Lustre.
Rather, the original intention of the author seems to be that
the test should always run with a single client. So, this patch
simply disables the test if there are several clients.

Change-Id: I98cec2388a0801e3bd0ee335ca4387a29cd35e35
Signed-off-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/7873
Reviewed-by: Alexander Nikolaevich Boyko <alexander.boyko@seagate.com>
Tested-by: Jenkins
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>