LUS-2578 mdt: Convert MDS restoring RPC message to D_WARNING
Using D_WARNING instead of D_RPCTRACE causes the message to be both
logged in the Lustre DK logs and on the system console.  This patch
changes the MDS restore/replay debug message to use D_WARNING.

The issue in bug 795155 went undetected because restoring metadata
requests was not noticed because it was under the D_RPCTRACE flag.  A
restored/replayed metadata request indicates some sort of underlying
error, and even when handled correctly, should generate a warning.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: I0da71d7c7435889f4dd29c91ff9cc2001a715b13
Reviewed-on: http://es-gerrit.dev.cray.com:8080/17258
Reviewed-by: Justin Miller <jmiller@cray.com>
Reviewed-by: Alexander Boyko <c17825@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>