MRP-3501 ofed: strengthen ib_create_cq test in 4.2+ kernels
External infinband stacks tend to keep their kernel
headers in sync with the latest kernels but this is
not the case with the driver code itself. Since this
is the case we can not trust testing for API changes
using OFED headers but need to be test driver code
API changes. Update the autoconf test to see if
ib_create_cq() uses struct ib_cq_init_attr.

Change-Id: I9875b6787723589b9cbdb55ef088b146b6900b2a
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: http://review.whamcloud.com/16342
Tested-by: Jenkins
Intel-bug-id: LU-6215
Reviewed-by: Alexander Boyko <alexander.boyko@seagate.com>
Reviewed-by: frank zago <fzago@cray.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/10995
Reviewed-by: Alexander Nikolaevich Boyko <alexander.boyko@seagate.com>
Tested-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>