MRP-3042 mdt: cancel granted lock on resend intent error
a race between intent and unlink results in working on orphan object,
the intent resend finds the granted lock and !existent object. Due to
a returned error on resend, lock is to be destroyed, what gets LBUG as
it is to be cancelled first.

It is unclear if the lock reached the client in the original reply,
just leave the lock on the server, not returning it again to the
client. Due to LU-6529, the server will not OOM in case the 2nd reply
will be handled on the client.

Signed-off-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Change-Id: I52c34f9718f056b0207a2a806c0f29727b287d20
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/8849
Tested-by: Jenkins
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>
Reviewed-by: Andriy Skulysh <andriy.skulysh@seagate.com>
Reviewed-by: Alexander Nikolaevich Boyko <alexander.boyko@seagate.com>