MRP-3137 xattr: ldlm_lock_decref_and_cancel()) ASSERTION
xattr code implies intent error is returned under a lock,
which is immediately cancelled. MRP-3072 stopped granting
a lock and the client failed as a lock is expected.

make all intent ops to work through md_intent_lock: getxattr
and layout, which should extract the intent error.

Check if a lock obtained and cancel it properly for error
cases. Note: we should support both cases for interop needs,
an intent error under a lock and with a lock abort.

Change-Id: Ic8d415a3f00758820dad34e1fea928b945345619
Signed-off-by: Andriy Skulysh <andriy.skulysh@seagate.com>
Signed-off-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/9166
Tested-by: Jenkins
Reviewed-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>