MRP-2739  Modified ll_find_alias to avoid cache corruption
ll_find_alias is responsible for getting alias for inode
which can be reused. Directories are asumed to have unique
alias. Where in case of non-directories there can be multiple
aliases. In case of lustre there can be two type of aliases
i.e. discon_alias and invalid_alias. Usage of discon_alias in
case of non-directories may corrupt dcache and leads to kernel
crash.Changes made to avoid use of discon_alias in case of
non-directories.

Change-Id: I3c36bf4437da71d79cd5dfe9d15a21aff24165ad
Signed-off-by: Lokesh Nagappa Jaliminche <lokesh.jaliminche@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/10304
Reviewed-by: Ujjwal Lanjewar <ujjwal.lanjewar@seagate.com>
Reviewed-by: Ashish Purkar <ashish.purkar@seagate.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Tested-by: Jenkins
Tested-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>