MRP-3196 test: properly export a temporary variable
$listvar is a list of all the ost-pools created. Whenever a new
ost-pool is created it is added to the list. Cleanup is done
based on the pools in the list.
If the check for empty list returns true it indicates there are
no pools to destroy and the script exits.

'eval export temp', which was added in patch for MRP-2171, does not export
$listvar properly thereby leading the check for empty list to return
true though the pool actually exists.

This patch makes the $listvar to be exported properly leading to
the expected cleanup at the end.

Signed-off-by: Bhagyesh Dudhediya <bhagyesh.dudhediya@seagate.com>
Change-Id: I445070f28fe7be5bbfecbc4c42d217f84a332743
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/9844
Reviewed-by: Ujjwal Lanjewar <ujjwal.lanjewar@seagate.com>
Reviewed-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>
Tested-by: Jenkins
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>