MRP-3249 ptlrpc: do not reduce replay request deadline
New timeout for a replay request sent to a client with early reply is
calculated without taking into account that new deadline for the
replay request on server side is calculated based upon
obd_recovery_timeout. That results in reducing replay deadline which
forces to early disconnect of the client.

Calculate the timeout sent in early reply to a replay request in
according to new deadline of the replay request on server side.

Signed-off-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Signed-off-by: Alexander Boyko <alexander.boyko@seagate.com>
Change-Id: I31e45e9906b044ba158ce0805d427dafa8292c12
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/10598
Tested-by: Jenkins
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>