MRP-4103 tests: skip tests when journal size is insufficient
Test 18a in recovery-small failed with invalid argument due to
journal size. When journal size is insufficient for given OST count,
not all the OST's can be be added in pool. In this case, the default
journal size is insufficient for 2 OSTs, so when request is placed
for 2nd OST, lod alloc_specific() returns EINVAL. This is because
"osts->op_count == 1 and lsm->lsm_oinfo[0]->loi_ost_idx == 1".
Due to this, client gets EINVAL when it sends set stripe in
mdc_intent_lock. Hence, Skipping tests when journal size is not
sufficient.

Signed-off-by: Jadhav Vikram <jadhav.vikram@seagate.com>
Change-Id: I53deca3d9cc0a56fc9ff5c4e507e22692cc69d2c
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/11101
Reviewed-by: Ashish Purkar <ashish.purkar@seagate.com>
Reviewed-by: Alexandr Boyko <alexander.boyko@seagate.com>
Tested-by: Parinay Vijayprakash Kondekar <parinay.kondekar@seagate.com>