MRP-3615 lod: lfs setstripe fix for pool
If a file is created (with lfs) in the directory associated
with pool without -p pool_name option then limit stripe count
to number of osts in the pool as that directory is associated
with the pool. This patch fixes this problem.

Also removed the wrong check from ost-pools.sh, test_20 where
we were creating file in a directory associated with pool and
checking it as not part of the pool.

Signed-off-by: Rahul Deshmukh <rahul.deshmukh@seagate.com>
Change-Id: I0671515dafecbd49465b2f3b9fb82e14c4663234
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/11274
Tested-by: Jenkins
Reviewed-by: Alexander Nikolaevich Boyko <alexander.boyko@seagate.com>
Reviewed-by: Vinayakswami Hariharmath <vinayakswami.hariharmath@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>