MRP-2960 llite: Implement lock ahead
This is the server side of the lock ahead feature.
The client side of this feature is not compatible with
Lustre 2.5.

The server side patch has the following components:
New LDLM flag, LDLM_FL_NO_EXPANSION, telling the server not
to expand extent locks requests with this flag set.
Changes to lock matching code to allow non-blocking lock
requests for all lock types (not just group locks).
Changes to the ofd intent code to get correct file sizes
with locks in advance of IO.

Additionally, this version of the patch has a module
parameter for the 'ofd' module, which turns lockahead
support on and off.  When the module parameter is not set
or is set to 0, the OSS will not send the OBD_CONNECT
flag for lock ahead, and the modified ofd intent/glimpse
code is not used.

Module parameter setting example, to enable lock ahead:
In a new file, /etc/modprobe.d/ofd.conf:
options ofd ofd_lockahead_enabled=1

The description of the full feature (including client code)
is here:
http://review.whamcloud.com/#/c/13564/

Signed-off-by: Patrick Farrell <paf@cray.com>
Signed-off-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Change-Id: I642d120f3fa10c500b1f6f96eefb25b81930efd8
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Reviewed-by: Alexey Leonidovich Lyashkov <alexey.lyashkov@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/11887/
Tested-by: Alexander Lezhoev <alexander.lezhoev@seagate.com>